Language System update
From now on, all language files will be stored as .lang, which is just
an Xml file. Only class that will remain is HPStrings, which is pattern
class. Languages will be copied to $conf$/lang on the first start and
after that, everything will be read from that directory. Language is
loaded in MainClass, before GUI is invoked and in HPSklad str is set to
HPStrings by default. Implementation into GUI will be done after
GraphicCore is done and ready.
Added LanguageLoader, which handles all actions regarding languages.
New package hexapaper.languages, where all the .lang are stored.
Removed commented old code from Config and mainClass